moved encounters up
