+++ app/src/actions/workspace.js
@@ -15,7 +15,7 @@ import { saveFilterGroup } from 'filters/filterGroupsActions';
 import { setOuterTimelineDates, SET_INNER_TIMELINE_DATES_FROM_WORKSPACE, setSpeed } from 'filters/filtersActions';
 import { setPinnedVessels, addVessel } from 'actions/vesselInfo';
 import { loadRecentVesselsList } from 'recentVessels/recentVesselsActions';
-import { setEncountersInfo } from 'mapPanels/rightControlPanel/actions/encounters';
+import { setEncountersInfo } from 'encounters/encountersActions';
 import calculateLayerId from 'util/calculateLayerId';
 import { hexToHue, hueToClosestColor } from 'util/colors';
 import uniq from 'lodash/uniq';
