+++ app/src/encounters/components/EncountersInfo.jsx
