+++ app/src/encounters/components/EncountersPanel.jsx
