+++ app/src/encounters/components/EncountersVessel.jsx
