+++ app/src/mapPanels/rightControlPanel/components/ControlPanel.jsx
@@ -11,7 +11,7 @@ import LayerPanel from 'layers/containers/LayerPanel';
 import LayerManagement from 'layers/containers/LayerManagement';
 import SearchPanel from 'search/containers/SearchPanel';
 import VesselInfoPanel from 'containers/Map/VesselInfoPanel';
-import EncountersPanel from 'mapPanels/rightControlPanel/containers/EncountersPanel';
+import EncountersPanel from 'encounters/containers/EncountersPanel';
 import ControlPanelStyles from 'styles/components/control_panel.scss';
 import iconStyles from 'styles/icons.scss';
 import SearchIcon from '-!babel-loader!svg-react-loader!assets/icons/search.svg?name=SearchIcon';
